Add initial version of microsoft-teams data acquisition
