+++ .gitignore
@@ -25,3 +25,5 @@ yarn-error.log*
 .idea
 
 .next
+
+*.sw[mnop]
