+++ package.json
@@ -26,6 +26,7 @@
     "koa-router": "8.0.8",
     "lodash": "4.17.15",
     "luxon": "1.22.2",
+    "msal": "1.3.1",
     "nanoid": "3.0.2",
     "next": "9.3.3",
     "next-images": "1.4.0",
