+++ pages/microsoft-teams/actions/index.ts
@@ -1,0 +1,3 @@
+import authorize from './authorize'
+
+export { authorize }
